Attached the File/Add menu to the Series/Import/Images... menu.
This allows the menus to operate somewhat similarly to the current
version. However, multiple files are not opened. This version has no
sense of a "series" yet. It just opens a single file.

The current Java prototype is being made to look as much like the
original Reconstruct as possible to make it familiar to the current user
base. It's a good place to start.