Added "Use" note for "Center Drawing" mode in documentation.
As mentioned in the documentation, the "Center Drawing" mode acts like a
table tool (jig saw or sewing machine) where the material is moved with
respect to a fixed work head. This can seem natural for some people, but
may also feel reversed for others. As suggested in the previous commit,
it might be nice (although difficult to implement) to reverse the
movement so it appears that the work head is being moved rather than the
material. This might provide a more natural feel. However, if the mouse
cursor is still visible, it would be moving in the OPPOSITE direction of
the movement of the material. This results from effectively having two
cursors: one for the mouse (following the mouse), and one for the point
of drawing (typically at the center). This would likely be confusing,
and so it might be desired to hide the mouse (give it an empty cursor)
in this mode. This would mean that the actual location of the mouse
would be unknown and may be difficult to "refind" when needed for other
operations (menu selections, window movement, etc).