Moved all of the previous top-level files into the source directory.
It's often convenient to use GitHub to manage more than just source
code. For example, documentation, notes, and discussions can all be
maintained in a GitHub repository. To make these different parts easier
to find, it's helpful to keep the source files (which tend to be many)
in a separate directory (often named "source" or "src").

This commit simply moved most of the top level files into a new "source"
directory to better separate them from other non-source files.