Added all of the GBM (Generalized Bit Map) files ... may need for TIFF
The original Reconstruct relied on the Generalized Bit Map code for most
of its image I/O. Java handles almost all formats internally except
possibly TIFF. The "C" code versions of these files might be helpful in
implementing a TIFF reading capability for the Java version.

Note that Java 9 does seem to have support for reading TIFF files, but
this is a relatively new capability.

Also note that one possible Java implementation would use Java for the
"front end" and use compiled C code for the "heavy lifting" portions of
Reconstruct. If this is done, then the TIFF reading code from the GBM
library might not be much of an additional burden - although it might
need to be integrated into the Java code while the other programs could
be external stand-alone programs.

Finally, it might be possible to precompile several different versions
of this special "C" code and include all of them in the .jar file. The
proper version to actually run could be selected at run time by the Java
version of Reconstruct (which will know the operating system in use).