Added support for multiple images as a plain Java array.
Added code to allow importing of multiple images via the menu item:

  Series / Import / Images...

In the past, this had imported a single image into the image_frame. The
previous image was lost. This commit adds an array of image_frames to
hold multiple images. Images are still added one at a time through the
same menu function, but the previous images are preserved in the array.

This version has no provision to select multiple images or to remove
images from the list (other than restarting the program). This version
also does not associate the strokes with any particular frame. There is
still a single set of strokes for the entire series.

Note also that the plain Java array might be better off as an ArrayList
in future versions. This version is primarily an exploration of the
performance of loading and switching between multiple images.