Python: Image scaling works, but can be slow for high zoom.
The current version essentially scales the entire image before drawing
it. This can create a huge image when zooming in - even though only a
small portion is currently showing.

This version uses the "scale_simple" method. The "scale" method might be
more efficient, and it's possible that there are other newer drawing
methods that could be faster as well.

For now, this seems to handle zooming and panning of the image.