Changed version to 1.2.0.0
This version is nearly identical to the previous version,
but the "1.2" reflects the start of a new series of changes.

The current versioning plan calls for rapid increases in the
least significant position (.0, .1, .8, .9, .10, .11, .352).
Important milestones will be reflected in the second least
significant position, and major releases will be reflected
in the top two (leftmost) positions.