Updated makefiles to build the gbm library as needed in ./gbm
The GBM library is needed by Reconstruct, and it makes sense (at least
for now) to maintain a copy of the source with Reconstruct and to build
it along with Reconstruct.

This commit modifies makefile_raw to also make the gbm library.

Note that executables and libraries are currently being maintained in
the repository. This isn't best practice, but it currently serves the
purposes of distributing the executable and tracking past builds that
may have performed differently from the current build.