Migrating from reconstruct_min_ok.rc to reconstruct.rc (1.2.0.2)
During the initial recompiling of Reconstruct, the resource file (.rc)
was built up from a minimal version by adding each section. That version
was named "reconstruct_min_ok.rc" as reflected in the earlier makefiles.

This commit brings those changes back into the original reconstruct.rc
file and renames the "min_ok" version to be "old". The "old" version
will be dropped from the repository in a future commit.

This is now version 1.2.0.2.