Added code to use help files (.chm) in location of Reconstruct.exe
This change gets the path to the current executable and then subtracts
off the length of the expected name "/Reconstruct.exe" (16 chars).

This is a quick fix to find the help files, but it should probably do a
better job of figuring out the path from the executable.

This is now version 1.2.0.4