Added .gitattributes file to specify binary and text files
Working with Windows has caused a lot of CR/LF issues. Many (if not all)
GIT documentation recommends that files in the database be stored with
linefeed (LF) endings. Then the settings on each machine can be used to
determine whether to convert to other formats (like CR/LF).

This commit adds a .gitattributes file to specify how the endings should
be handled for each file type or individual file name.

In addition to the .gitignore file, the following should be set:

  Linux (LF) Machines:
    git config core.autocrlf input

  Windows (CRLF) Machines:
    git config core.autocrlf true

This means that Linux machines will convert CR/LF to LF when files are
stored in the .git database ("input" to the database). Windows machines
will automatically convert between CR/LF in the working areas and LF in
the .git database.

Note that this is still being tested, and may need to be changed.

Also note that the current database does have some CR/LF corruption
which should be removed.

References:
  https://help.github.com/articles/dealing-with-line-endings
  http://adaptivepatchwork.com/2012/03/01/mind-the-end-of-your-line/
  https://git-scm.com/book/en/v2
  http://mike.meessen.biz/blog/?p=368
  http://opendirective.net/blog/2015/12/dealing-with-windows-text-line-endings-in-git/