Normalize line endings to be LF in the databse per GitHub article
From:  https://help.github.com/articles/dealing-with-line-endings

This source (and others) recommends using LF in the GIT database and
having machines configured to convert as needed.

  Linux machines should always convert CRLF to LF when committing
  Windows machines should always use auto CRLF to convert both ways

This requires normalization of the Git database to remove CRs as
described in the previous link. These steps were taken:

    git commit
    rm .git/index
    git status
    git reset
    git status
    git add -u
    git add .gitattributes
    ls -al | grep git
    git add .gitignore
    git status
    git commit

That appears to have modified almost all of the source files.