Added a Python3 GTK example with XML Menu Definition
The python_gtk version is based on an earlier version of GTK. It's poor
image performance might be related to the image handling capabilities of
that version (or possibly the implementation).

This version attempts to use the newer GTK version and is patterned
after an example found at:

 https://developer.gnome.org/gnome-devel-demos/stable/menubar.py.html.en

This version contains an internal XML file, but it could be read from a
file just as well (future versions may support this).