Feature: Hide/Unhide all traces without picking a domain (1.2.0.5)
This is the first step in providing a key to hide/unhide all traces.
This was submitted as an item in the Wish List (Wish_List.md).

This commit does not add the feature, but simply adds a new key to use
for testing. The chosen key was "Control-E" since it's the first unused
key in the Reconstruct key list. It could be changed later.

To test the mapping of the new "Control-E" key, this version simply
exits whenever "Control-E" is pressed. An attempt was made to print
using "printf", but that failed to produce output. This should be
resolved since printing can be a helpful debugging tool.

Note that this version change (from 1.2.0.4 to 1.2.0.5) reflects an
incompleted feature. It's not clear how to best represent that in our
version numbering system. One approach would be to use odd numbers for
features that are not complete and only move to even numbers when the
feature is working. Alternatively, any non-zero lowest number could be
considered as a "work in progress" and the feature is only considered
complete when the lowest number is 0. Note that any use of the lowest
number conflicts with its previous use of signifying 32=bit (0) and
64-bit (1) versions.