1.2.0.6: Attempted hide/unhide contours with the Control-E key (not working)
This version intercepts the Control-E key and attempts to hide/unhide
all of the contours by setting their "hidden" flag.

The main logic is in keyboard.cpp / CmHideAllTraces(). The logic
searches through the "CurrSeries->contours list looking for any that are
visible (not hidden). It prints whether any are visible or not. Then it
sets all contours to be hidden if any were visible and to visible if all
were hidden. The printed report seems to toggle for successive Control-E
presses, however, the contours remain visible.

It appears that contour visibility requires further investigation.

As far as the version numbering, there's a good argument that every
"push" to the repository should have a different version number since
every "push" is a version that might be used by someone. Intermediate
versions (batched within the same push) may have the same number, but
anyone wanting to use such an older version would have to explicitly
check out that older version. In other words, it wouldn't be what they
would get if they just "pulled" from the repository. For that reason,
this commit is version 1.2.0.6 while all intermediate commits before
this one are still 1.2.0.5. This suggests that any commits that don't
end in a zero are "development" versions. They are being developed in
the "master" branch for now, but in the future such development versions
could be kept in a "development" branch of the Git repository. That's
not being done at this time because it's easier for people who aren't
familiar with Git (and GitHub) to just pull from the "master" branch,
and we're trying to encourage as much testing as possible.