PyGtk: Commented attempts to scale only part of the image.
The PyGtk version (not PyGtk3) scales the entire image before drawing
it. This can produce a very large image of which only a small portion
will be drawn in the window.

This version uses "scale_simple" which is convenient because it
generates the resulting image of the right type (bits, alpha, etc).

The more general "scale" function requires a destination image which
must be created in advance. This became a problem since the Pixbuf
constructor fails with 24 bit images and only seems to work with 8.
These limitations forced the abandonment of the approach, but the
partially completed code remains as a comment.