PyGTK3: More attempts to draw images with cairo / PyGTK 3
Several of these attempts have resulted in a locked window manager.

It's not clear what's going on, but this code may be mixing different
mechanisms of the Python GTK system (PyGObject, GTK, gi, cairo). It may
be that they're not all compatible.

This version has all of the troublesome code commented out, so it
doesn't crash, but it doesn't draw images either.