Added basic parsing of section files (in addition to series files).
Section files are not explicitly listed in the series file. However,
all files with the same "first name" (before the trailing dot) as the
series file are considered to be the sections that make up the series.

These section files are numbered from 1 up to the number of sections in
the series. Their filename extension is that number (in decimal) with no
leading zeros ("first_name.1", "first_name.2", ... "first_name.123").

The section files are also XML (even though they have a numeric extension)
and they contain an "index" attribute field (such as index="9") that
should presumably be the same number as the file name extension. If
these should differ, it's not clear which would take precedence.

This commit provides code to parse each of those section files and prints
the results to the screen. It starts by getting the first name from the
series and then filtering all other files in that directory to match the
series name followed by "." followed by a number. Then each section is
also parsed and the results are printed.