Moved much functionality from Reconstruct.java into SeriesClass.java.
The new SeriesClass constructor now builds its own document object from
a series file. It also currently builds the sections (which should
ultimately be done in the SectionClass). This removes that functionality
from the main Reconstruct.java class.

This version still has the data stored in document objects (as it came
from the XML). It's not clear if that's the best storage method or if
something like a HashMap might be more appropriate.