Reversed a few of the features of drawing mode and movement mode.
In movement mode: scroll wheel zooms, left mouse drag moves
In drawing mode: scroll wheel moves through series, left mouse draws

These seem more natural.

Also, this version was compiled with javac 1.7.0_171 in order to be
backward compatible with earlier Java Runtime Environments (JREs).
Previous (recent) versions had been compiled with javac 9.0.4 in order
to support ".tif" images. However, it turns out that it's the run time
environment (JRE) that determines the ability to load ".tif" files and
NOT the original compiler used to build the ".jar" file. This is because
the file types are loaded dynamically based on the file extension. So
when an older JRE is asked to load a ".tif" file, it fails because it
doesn't recognize that format. But when the same ".jar" file is run with
a newer JRE, it will look up the ".tif" extension and find that it can
read that file and return an image instance that the older JRE (and the
older ".jar" file) can handle.

However, there seem to be incompatibilities that prohibit older JRE's
from running ".jar" files compiled with the newer compiler. Since the
current Debian version used in the lab has JRE 1.7.0_171 as the default,
this code will continue to be compiled with JDK 1.7 as well. That code
will not be able to open ".tif" files when run under JRE 1.7, but it
will be able to open ".tif" files when it's run in the newer JRE.