Began moving drawing from Reconstruct to Sections.
The sections should really have the data to be drawn (images and
traces), so it makes sense for them to draw themselves.

This commit added a "paint_section" method to the SectionClass to
support drawing by the section. This code isn't called yet, and this
version is still doing all drawing from Reconstruct.java. But this is
the first step in moving to a more localized delegation model.