Disabled much of old direct image reading code (source only update).
Earlier versions were reading directly from a series of images. The last
few commits have been working toward reading "Sections" rather than
images. Each such "Section" is an instance of "SectionClass" which
contains its own image(s). At this point, each section is only expected
to contain one image ("domain"), but provisions are added to allow for
multiple images in the future.

This version still does the drawing from the series, but that may be
moved into the sections when the strokes are moved as well.