Added BezierTracing.java and .jar (uses Bezier Curves for tracing).
This is a relatively crude demonstration of an interface that may be
helpful for reconstruct.

This code was based on code originally written in QBASIC may years ago
and later ported to java's JDK 1.1 (also many years ago). Run with:

   java -jar BezierTracing.jar [image_file_name]

User interface:

  Right Mouse Button:  Starts and Stops a Trace
  Left Mouse Button:  Adds or moves points.

Typical workflow:

  1. Right click to start a trace (white dot appears)
  2. Left click to add sequential points (no dragging needed)
  3. Right click to end the trace (closing is based on location)
  4. Use normal (left) button to move points or handles

The curve will be closed or not depending on the mouse location
at the time of the closing "Right click" (step 3. above). If the
mouse location is closer to the first point drawn than it is to
the last point drawn, then the curve will be closed (from the last
point to the first point). But if the mouse location is closer to
the last point drawn than to the first point, then the curve will
be open.

Using this workflow, every click (except the last for open curves)
will generate either a point or a segment.

When not in the middle of drawing a curve (delimited with right
clicks), the left mouse button can select any point (or handle)
and drag to move.