Added "TraceRound.gif" to demonstrate the Experimental Java Interface
This experimental interface eliminates dragging when making an original
trace. Instead, mouse clicks are used to create Bezier curves.

The RIGHT mouse button is used to start and end a trace.

All other clicks are with the LEFT mouse button.

When not actually drawing, the left mouse button is used to select and
drag points (either end points or Bezier "handles") as is customary.

This commit contains an animated GIF  demonstrating how this is done.