Updated the makefile to generate SHA1 sum for BezierTracing.jar
Note that there are some problems with the current makefile that arise
from having multiple Java classes in a single source file.

In Java, each class is typically compiled to its own ".class" file even
when multiple classes share the same source (.java) file. This makes it
more difficult to use a simple make file because those classes may need
to be known explicitly so they can be bundled into a single .jar file.

The easiest approach is to simply compile all the java files and then
put all class files ("*.class") into a .jar file. But when the source
for two different .jar files share the same directory, the use of
*.class will pull in the class files intended for ALL .jar files.

One solution is to simply delete all the .class files at each step in
the makefile, but then this defeats one useful aspect of a makefile
(only compiling what's needed).

There may be ways around this, but they will likely complicate the
makefile. For now, it's best to "make clean; make".