Added an option (now default) to align handles (not perfect yet)
The "a" key will toggle the alignment of the slope handles through their
related point. The sizing of these handles is somewhat undefined at this
point in time. The original plan was to just change the angle and scale
them both, but then it's not clear how they would be made different
relative to one another.

But overall, this gives a feeling for how it would work, and it has been
used to generate a fairly complicated trace.

Also, new Dump and Load options were added as keys but not implemented.