Added code to save and load tracings in BezierTracing.
This adds a simple DataOutputStream and DataInputStream mechanism for
saving tracings to a file and loading them back again. The file name is
currently hard-coded to be "bezier_traces.dat", and there is no check on
over-writing previous files (copy to other names as needed). The data
files are a binary format reflecting the internal Java representations
of the values.