Added scaling and offsets to XML points to match image on screen.
The data trace points in the Reconstruct XML file appear to take values
from 0 to 1. It's not clear what they mean at this time. This change
applies a scale/offset to get the XML data to match what's being drawn
on the screen (for the "calibrated_organelle" series). This required a
scaling factor of 165 and an offset of 100 for both x and y. It's not
clear what conversion (or conversions) these numbers represent, but they
do seem to make the XML traces match the images at all scales and
offsets. More research is needed.