Modified calibrated_organelle.1 to work with "recon2obj" program.
For some reason, the previous version was not properly read by the
recon2obj program. That version had been hand edited to put the
transform coefficients on the same line as the Transform tag, and to
indent the contents of the transform (Image and Contour tags). The only
changes were "whitespace" changes, but they caused the previous version
to not work with "recon2obj". It may be that recon2obj is making some
assumptions about the actual format of the XML file it reads. This is
not unreasonable since these files are normally generated by Reconstruct
and would normally follow identical formatting. However, it might be
worth fixing for environments where the XML comes from other sources.