Added Cube_Slices and Cube_Ideal series for testing transforms.
These series were constructed from 16 simple images of cube slices. The
images were created as PNG files in a paint program and converted to
JPEG for use with the older Reconstruct (which doesn't read PNG files).

The images in the series alternate between untransformed and transformed
images of the same 200x200 square. The transforms are shifting in x and
y, scaling up and down, rotating +30 and -30 degrees, and a fairly
complex transform of the square to an irregular shape (retaining
corners).

These images were then traced in Reconstruct 1.2.0.6 to create a series
named "Cube_Slices.ser". A second series ("Cube_Ideal.ser") was created
from the first by simply copying the series and section files and
rounding the transform values to what would be the ideal transformations
to return to the original 200x200 squares. Some of these values were
guesses, and may benefit from further refinement.

These series files will serve to test the transforms needed to properly
interpret the Reconstruct XML files.