Added a hand-typed "section.dtd" taken from Reconstruct User Manual
The Reconstruct User Manual (reconstructusermanualv1.1.0.0.pdf) contains
the section.dtd file listed on pages 14 and 15. The PDF is "SECURED"
which seems to prevent copying of text, so this copy of the series.dtd
file was hand-typed while viewing the PDF file. It might need to be
checked for accuracy.

The manual (page 15) states:

  "Note that a similar DTD is defined for the series data file
   (series.dtd"), but it is not expounded here because the large
   number of series options makes the file very long."

At this point in time, the series.dtd is still missing from this source
repository.