Added out of memory checking and primitive purging system.
The purging system is in "SeriesClass.paint_section". Right now it
removes images from the "opposite side" of the circular image stack
until it's able to load the next image. Images are removed from the
opposite side and "toward" the current image in both directions.

This would be ideal if the view actually wrapped around (which it
doesn't currently do), but it was an easy start.

In practice, this method only removes 1 or 2 images at a time which
means that when memory is full, it will pause to purge on every image
change. It might be better to remove more images when memory is full
or even detect when it's getting full to keep the application more
responsive to scrolling. This may be revisited in the future.